bugfixing introduced by localization update.
Tetris aborted initialization because it was waiting for an element (score) to be avalaible. since its a number that updates, it wasnt avaliable. and broke the logout again + tetris.

This was fixed by altering the i18n data key with span score label, and leaving the updating score as a span score.

working now again